% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_Functions_for_analysis_MTG_Commander.R
\name{hypergeo_variance}
\alias{hypergeo_variance}
\title{hypergeometric variance}
\usage{
hypergeo_variance(N, K, n)
}
\arguments{
\item{N}{population size, K is the sample size, n the number of accesses}
}
\value{
Variance of a probability
}
\description{
hypergeometric variance
}
