% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_Functions_for_analysis_MTG_Commander.R
\name{Load_ScryfallCard}
\alias{Load_ScryfallCard}
\title{Function loads the JSON card data from Scryfall}
\usage{
Load_ScryfallCard(id)
}
\arguments{
\item{scryfall}{ID}
}
\value{
a list derived from the JSON given by the Scryfall API that represents a single card
}
\description{
Function loads the JSON card data from Scryfall
}
