% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_Functions_for_analysis_MTG_Commander.R
\name{probability_exact}
\alias{probability_exact}
\title{Exact hypergeometric probability}
\usage{
probability_exact(x, draws, y, deck_size)
}
\arguments{
\item{X, }{draws, y and deck size which is the population site}
}
\value{
A probabiity
}
\description{
Exact hypergeometric probability
}
