% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_Functions_for_analysis_MTG_Commander.R
\name{return_mv_count}
\alias{return_mv_count}
\title{Functions to return mana value distribution from individual values}
\usage{
return_mv_count(mana_values)
}
\arguments{
\item{vector}{of mana values}
}
\value{
A database of mana value counts
}
\description{
Functions to return mana value distribution from individual values
}
