% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_Functions_for_analysis_MTG_Commander.R
\name{run_Atleast_land_base_test}
\alias{run_Atleast_land_base_test}
\title{Hypergeoemtric data gathering function. At least so many probabilities}
\usage{
run_Atleast_land_base_test(
  At_least_x,
  max_number_in_deck,
  min_number_in_deck = 1,
  deck_size = 99,
  draws = 7
)
}
\arguments{
\item{X, }{max number in deck, min number in deck, deck or population size, draws or sample size}
}
\value{
A database of probabilities
}
\description{
Hypergeoemtric data gathering function. At least so many probabilities
}
